feat: Added SL and SFL protocols
Needed to apply some tricks to make it work. Most importantly, the way the data is sampled needs to be adapted using the classes in sl_distributed_datamodule, which works for both SL and SFL. It remains a small issue though: batches from following clients may overlap. This may be solved through a custom batch sampler.